requirements.txt
ollama
gymnasium[accept-rom-license]
stable-baselines3
pyyaml